backend/change-password/requirements.txt
passlib==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.9